+++ src/components/temput.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 function paste() {
   console.log("clicked!");
-  const text = "hogehoge";
-  const type = "text/plain";
+  const text = "<ul><li><p>aaa</p><ul><li><p><strong>bbb</strong></p></li></ul></li></ul>";
+  const type = "text/html";
   const blob = new Blob([text], { type });
   const data = [new ClipboardItem({ [type]: blob })];
   navigator.clipboard.write(data);
